♻️ Merge package scripts into `curler.py`
