--- curler/__init__.py
@@ -1,10 +1,0 @@
-"""
-`curler` is a lightweight package to parse curl commands to use them in python.
-
----
-
-Author: https://github.com/arv-anshul/
-Homepage: https://github.com/arv-anshul/curler/
-"""
-
-from curler.parser import parse_curl, parse_file
