--- curler/constants.py
@@ -1,6 +1,0 @@
-from typing import NamedTuple
-
-
-class Method(NamedTuple):
-    GET: str = "GET"
-    POST: str = "POST"
