+++ tests/test_parser.py
@@ -1,27 +1,35 @@
 import unittest
 
-from curler import typing
-from curler.parser import parse_curl
+from curler import parse_curl
 
 
 class TestCurlParser(unittest.TestCase):
     def test_get_request(self):
-        curl_command = "curl http://example.com"
-        parsed = parse_curl(curl_command)
-        self.assertEqual(parsed.method, typing.http_method.GET)
+        command = "curl http://example.com"
+        parsed = parse_curl(command)
+        self.assertEqual(parsed.method, "GET")
 
     def test_post_request(self):
-        curl_command = 'curl -X POST -d "data" http://example.com'
-        parsed = parse_curl(curl_command)
+        command = 'curl -X POST -d "data" http://example.com'
+        parsed = parse_curl(command)
         self.assertEqual(parsed.method, "POST")
 
+    def test_other_request_mthods(self):
+        # --- --- PUT METHOD --- --- #
+        parsed_put_method = parse_curl('curl -X PUT -d "data" http://example.com')
+        self.assertEqual(parsed_put_method.method, "PUT")
+
+        # --- --- DELETE METHOD --- --- #
+        parsed_delete_method = parse_curl('curl -X DELETE -d "data" http://example.com')
+        self.assertEqual(parsed_delete_method.method, "DELETE")
+
     def test_cookies_and_headers(self):
-        curl_command = """curl \
+        command = """curl \
             -H "Cookie: cookie1=value1; cookie2=value2" \
             -H "Authorization: Bearer token" \
             -H "Content-Type: application/json" \
             http://example.com"""
-        parsed = parse_curl(curl_command)
+        parsed = parse_curl(command)
         self.assertEqual(parsed.cookies, {"cookie1": "value1", "cookie2": "value2"})
         self.assertEqual(
             parsed.headers,
@@ -32,11 +40,11 @@ class TestCurlParser(unittest.TestCase):
         )
 
     def test_proxy_and_user_auth(self):
-        curl_command = """curl \
+        command = """curl \
             -x proxy.example.com:8080 \
             -U user:password \
             http://example.com"""
-        parsed = parse_curl(curl_command)
+        parsed = parse_curl(command)
         self.assertEqual(parsed.user, ("user", "password"))
         self.assertEqual(
             parsed.proxy,
@@ -45,10 +53,6 @@ class TestCurlParser(unittest.TestCase):
                 "https": "https://user:password@proxy.example.com:8080/",
             },
         )
-
-    def test_not_implemented_method(self):
-        curl_command = "curl -X PUT http://example.com"
-        self.assertRaises(NotImplementedError, parse_curl, curl_command)
 
     def test_for_requests_method(self):
         command = "curl https://example.com"
@@ -60,7 +64,7 @@ class TestCurlParser(unittest.TestCase):
             "Check ordering of the list elements.",
         )
         self.assertEqual(parsed["url"], "https://example.com")
-        self.assertEqual(parsed["method"], typing.http_method.GET)
+        self.assertEqual(parsed["method"], "GET")
 
 
 if __name__ == "__main__":
